requirements.txt
google-adk==1.14.0
python-dotenv
langchain-community
stackapi
